Removing mesh folder
